+++ frontend/src/components/scenario/ScenarioBox.tsx
@@ -3,7 +3,7 @@ import ScenarioGraph from "./ScenarioGraph";
 import TextInputBox from "../common/TextInputBox";
 import LinkButton from "../common/LinkButton";
 import ActionButton from "../common/ActionButton";
-import { useScenario } from "@/hooks/scenario/useScenario";
+import { useScenario } from "@/hooks/useScenario";
 import { useRouter } from "next/navigation";
 
 const ScenarioBox = ({
