+++ frontend/src/components/testcase/TestcaseBox.tsx
@@ -3,7 +3,7 @@ import TestcaseDataTable from "./TestcaseDataTable";
 import TextInputBox from "../common/TextInputBox";
 import LinkButton from "../common/LinkButton";
 import ActionButton from "../common/ActionButton";
-import { useTestcase } from "@/hooks/testcase/useTestcase";
+import { useTestcase } from "@/hooks/useTestcase";
 import { useRouter } from "next/navigation";
 
 export const TestcaseBox = ({
