+++ frontend/src/hooks/useProjectAuth.ts
