+++ frontend/src/hooks/useScenario.ts
